+++ Shop/Query/Orders/DTOs/OrderItemDto.cs
@@ -11,6 +11,6 @@ public class OrderItemDto : BaseDto
     public Guid OrderId { get; set; }
     public Guid InventoryId { get; set; }
     public int Count { get; set; }
-    public long Price { get; set; }
-    public long TotalPrice => Price * Count;
+    public int Price { get; set; }
+    public int TotalPrice => Price * Count;
 }
